refactor(api/games): split save_game into persist.rs, item sync into items.rs
