+++ api/src/games/mod.rs
@@ -1,5 +1,7 @@
 pub mod handlers;
 pub use handlers::*;
+pub(crate) mod items;
+pub(crate) mod persist;
 
 use crate::tributes::TRIBUTES_ROUTER;
 use crate::{AppError, AppState};
@@ -350,7 +352,7 @@ async fn run_game_cycles(
 ) -> Result<(), AppError> {
     game.run_full_day()
         .map_err(|e| AppError::InternalServerError(format!("Failed to run game day: {}", e)))?;
-    let _ = save_game(game, db, broadcaster).await?;
+    let _ = persist::save_game(game, db, broadcaster).await?;
     // Persistence has happened; the cycle's `CycleStart`/`CycleEnd`
     // GameMessages were broadcast by `save_game` as it drained
     // `game.messages`, so the frontend's `GamePage` already invalidates
@@ -429,419 +431,4 @@ impl From<GameLog> for GameMessage {
             payload,
         }
     }
-}
-
-async fn save_game(
-    game: &mut Game,
-    db: &Surreal<Any>,
-    broadcaster: &crate::websocket::GameBroadcaster,
-) -> Result<Json<Game>, AppError> {
-    let game_identifier = RecordId::from(("game", game.identifier.clone()));
-
-    // Start transaction
-    db.query("BEGIN TRANSACTION").await.map_err(|e| {
-        AppError::InternalServerError(format!("Failed to start transaction: {}", e))
-    })?;
-
-    // Drain events accumulated during the most recent run_day_night_cycle.
-    // Persist them to the message table and broadcast to subscribed WS clients.
-    let logs: Vec<GameMessage> = std::mem::take(&mut game.messages);
-    if !logs.is_empty() {
-        let game_day = game.day.unwrap_or_default();
-
-        // Broadcast first so clients see updates even if persistence is slow.
-        for log in &logs {
-            crate::websocket::broadcast_game_message(broadcaster, &game.identifier, log.clone());
-        }
-
-        let game_logs: Vec<GameLog> = logs
-            .into_iter()
-            .map(|log| GameLog {
-                id: RecordId::from(("message", &log.identifier)),
-                identifier: log.identifier,
-                source: log.source,
-                game_day,
-                subject: log.subject,
-                timestamp: log.timestamp,
-                content: log.content,
-                phase: log.phase,
-                tick: log.tick,
-                emit_index: log.emit_index,
-                payload: serde_json::to_string(&log.payload).unwrap_or_else(|_| "null".to_string()),
-            })
-            .collect();
-
-        // Check current message count for this game
-        let current_count: u32 = db
-            .query(
-                "RETURN count(SELECT id FROM message WHERE string::starts_with(subject, $game_id))",
-            )
-            .bind(("game_id", game.identifier.clone()))
-            .await
-            .ok()
-            .and_then(|mut r| r.take(0).ok().flatten())
-            .unwrap_or(0);
-
-        let new_messages_count = game_logs.len();
-        let total_after_insert = current_count as usize + new_messages_count;
-
-        // Log warning if approaching limit
-        if current_count >= 9000 && current_count < MAX_MESSAGES as u32 {
-            tracing::warn!(
-                game_id = %game.identifier,
-                current_count = %current_count,
-                "Game message count approaching limit (9000+)"
-            );
-        }
-
-        // Rotate messages if total would exceed MAX_MESSAGES
-        if total_after_insert > MAX_MESSAGES {
-            let messages_to_delete = total_after_insert - MAX_MESSAGES;
-            tracing::info!(
-                game_id = %game.identifier,
-                current_count = %current_count,
-                new_count = %new_messages_count,
-                deleting = %messages_to_delete,
-                "Rotating old messages to maintain {} message limit", MAX_MESSAGES
-            );
-
-            if let Err(e) = db
-                .query(
-                    r#"
-                    DELETE message
-                    WHERE string::starts_with(subject, $game_id)
-                    ORDER BY timestamp ASC
-                    LIMIT $delete_count
-                    "#,
-                )
-                .bind(("game_id", game.identifier.clone()))
-                .bind(("delete_count", messages_to_delete))
-                .await
-            {
-                tracing::error!(
-                    game_id = %game.identifier,
-                    error = %e,
-                    "Failed to rotate old messages"
-                );
-                // Continue anyway - this is not critical enough to fail the save
-            }
-        }
-
-        // The structured `event` payload rides as a plain JSON string
-        // on the `GameLog.event` field, sidestepping the SurrealDB SDK
-        // serializer's habit of collapsing externally-tagged enums and
-        // `serde_json::Value::Object` payloads to `{}` when bound into
-        // an `object` column. mqi.3.
-        if let Err(e) = db.insert::<Vec<GameLog>>(()).content(game_logs).await {
-            let _ = db.query("ROLLBACK").await;
-            return Err(AppError::InternalServerError(format!(
-                "Failed to save game logs: {}",
-                e
-            )));
-        }
-    }
-
-    let area_results = futures::future::join_all(game.areas.iter().map(|area| async {
-        let id = RecordId::from(("area", area.identifier.clone()));
-        save_area_items(&area.items, id.clone(), db).await?;
-
-        let mut area_without_items = area.clone();
-        area_without_items.items = vec![];
-        // Bind via serde_json::Value to bypass the SurrealDB SDK's bespoke
-        // type serializer, which collapses externally-tagged enums and Option
-        // fields. The generic JSON bind path round-trips cleanly.
-        let body = serde_json::to_value(&area_without_items)
-            .map_err(|e| AppError::InternalServerError(format!("Failed to encode area: {}", e)))?;
-        db.query("UPSERT $rid CONTENT $body")
-            .bind(("rid", id.clone()))
-            .bind(("body", body))
-            .await
-            .map_err(|e| AppError::InternalServerError(format!("Failed to update area: {}", e)))
-            .map(|_| ())
-    }))
-    .await;
-
-    if let Some(err) = area_results.iter().find_map(|r| r.as_ref().err()) {
-        let msg = format!("{}", err);
-        let _ = db.query("ROLLBACK").await;
-        return Err(AppError::InternalServerError(format!(
-            "Failed to save area items: {}",
-            msg
-        )));
-    }
-
-    let tribute_results = futures::future::join_all(game.tributes.iter().map(|tribute| async {
-        let id = RecordId::from(("tribute", tribute.identifier.clone()));
-        if tribute.is_alive() {
-            save_tribute_items(&tribute.items, id.clone(), db).await?;
-        }
-
-        let mut tribute_without_items = tribute.clone();
-        tribute_without_items.items = vec![];
-        // Same workaround as the area UPDATE above: serde_json::Value bypasses
-        // the SDK's enum-collapsing serializer.
-        let body = serde_json::to_value(&tribute_without_items).map_err(|e| {
-            AppError::InternalServerError(format!("Failed to encode tribute: {}", e))
-        })?;
-        db.query("UPSERT $rid CONTENT $body")
-            .bind(("rid", id.clone()))
-            .bind(("body", body))
-            .await
-            .map_err(|e| AppError::InternalServerError(format!("Failed to update tribute: {}", e)))
-            .map(|_| ())
-    }))
-    .await;
-
-    if let Some(err) = tribute_results.iter().find_map(|r| r.as_ref().err()) {
-        let msg = format!("{}", err);
-        let _ = db.query("ROLLBACK").await;
-        return Err(AppError::InternalServerError(format!(
-            "Failed to save tribute items: {}",
-            msg
-        )));
-    }
-
-    // Persist mutable game fields explicitly. `db.update().content()` with the
-    // SurrealDB SDK's bespoke serializer drops `Option<u32>` fields like `day`
-    // (same family of bugs as the externally-tagged-enum collapse called out
-    // around the message payload above), so we use a plain UPDATE query that
-    // names the fields we want written.
-    if let Err(e) = db
-        .query("UPDATE $record_id SET day = $day, status = $status")
-        .bind(("record_id", game_identifier.clone()))
-        .bind(("day", game.day))
-        .bind(("status", game.status.to_string()))
-        .await
-    {
-        let _ = db.query("ROLLBACK").await;
-        return Err(AppError::InternalServerError(format!(
-            "Failed to update game: {}",
-            e
-        )));
-    }
-
-    db.query("COMMIT").await.map_err(|e| {
-        AppError::InternalServerError(format!("Failed to commit transaction: {}", e))
-    })?;
-    Ok(Json(game.clone()))
-}
-
-async fn save_area_items(
-    items: &Vec<Item>,
-    owner: RecordId,
-    db: &Surreal<Any>,
-) -> Result<(), AppError> {
-    // Get existing items
-    let existing_items: Vec<Item> = db
-        .query("SELECT * FROM items WHERE in = $owner")
-        .bind(("owner", owner.clone()))
-        .await
-        .map_err(|e| AppError::InternalServerError(format!("Failed to fetch items: {}", e)))?
-        .take(0)
-        .map_err(|e| AppError::InternalServerError(format!("Failed to take items: {}", e)))?;
-
-    // Create lookups for efficient comparison
-    let mut existing_map = HashMap::new();
-    for item in existing_items {
-        existing_map.insert(item.identifier.clone(), item.clone());
-    }
-
-    let mut new_map = HashMap::new();
-    for item in items {
-        new_map.insert(item.identifier.clone(), item.clone());
-    }
-
-    // Find items to delete (in DB but not in new items or durability is 0)
-    let mut items_to_delete = Vec::new();
-    for id in existing_map.keys() {
-        if let Some(item) = new_map.get(id) {
-            if item.current_durability == 0 {
-                items_to_delete.push(id.clone());
-            }
-        } else {
-            items_to_delete.push(id.clone());
-        }
-    }
-
-    // Find items to update (in DB and in new items with different values)
-    let mut items_to_update = Vec::new();
-    for (id, item) in &new_map {
-        if item.current_durability > 0 {
-            if let Some(existing) = existing_map.get(id) {
-                if existing != item {
-                    items_to_update.push(item.clone());
-                }
-            } else {
-                items_to_update.push(item.clone());
-            }
-        }
-    }
-
-    // Update relations - first delete existing relations
-    db.query("DELETE FROM items WHERE in = $owner")
-        .bind(("owner", owner.clone()))
-        .await
-        .map_err(|e| AppError::InternalServerError(format!("Failed to delete items: {}", e)))?;
-
-    // Batch delete operations
-    if !items_to_delete.is_empty() {
-        let delete_ids: Vec<String> = items_to_delete
-            .iter()
-            .map(|id| format!("item:{}", id))
-            .collect();
-
-        db.query("DELETE item WHERE id IN $ids")
-            .bind(("ids", delete_ids))
-            .await
-            .map_err(|e| {
-                AppError::InternalServerError(format!("Failed to batch delete items: {}", e))
-            })?;
-    }
-
-    // Batch update operations
-    if !items_to_update.is_empty() {
-        // Use serde_json::to_value + bound CONTENT so ALL Item fields round-trip
-        // (the previous hand-rolled CONTENT block silently dropped `rarity`,
-        // and string-interpolating fields is fragile around quoting).
-        for item in &items_to_update {
-            let rid = RecordId::from(("item", item.identifier.to_string().as_str()));
-            let body = serde_json::to_value(item).map_err(|e| {
-                AppError::InternalServerError(format!("Failed to encode item: {}", e))
-            })?;
-            db.query("UPSERT $rid CONTENT $body")
-                .bind(("rid", rid))
-                .bind(("body", body))
-                .await
-                .map_err(|e| {
-                    AppError::InternalServerError(format!("Failed to update item: {}", e))
-                })?;
-        }
-
-        // Batch insert relations. Hyphenated UUIDs must be wrapped in
-        // ⟨angle brackets⟩ or Surreal's SQL parser splits them on `-`.
-        let mut relation_parts = Vec::new();
-        for item in &items_to_update {
-            relation_parts.push(format!(
-                "RELATE {}->items->item:⟨{}⟩",
-                owner, item.identifier
-            ));
-        }
-
-        let bulk_relations = relation_parts.join(";\n");
-        db.query(&bulk_relations).await.map_err(|e| {
-            AppError::InternalServerError(format!("Failed to batch create relations: {}", e))
-        })?;
-    }
-
-    Ok(())
-}
-
-async fn save_tribute_items(
-    items: &Vec<Item>,
-    owner: RecordId,
-    db: &Surreal<Any>,
-) -> Result<(), AppError> {
-    // Get existing items
-    let existing_items: Vec<Item> = db
-        .query("SELECT * from owns->items WHERE in = $owner")
-        .bind(("owner", owner.clone()))
-        .await
-        .map_err(|e| AppError::InternalServerError(format!("Failed to fetch items: {}", e)))?
-        .take(0)
-        .map_err(|e| AppError::InternalServerError(format!("Failed to take items: {}", e)))?;
-
-    // Create lookups for efficient comparison
-    let mut existing_map = HashMap::new();
-    for item in existing_items {
-        existing_map.insert(item.identifier.clone(), item);
-    }
-
-    let mut new_map = HashMap::new();
-    for item in items {
-        new_map.insert(item.identifier.clone(), item.clone());
-    }
-
-    // Find items to delete (in DB but not in new items or durability is 0)
-    let mut items_to_delete = Vec::new();
-    for id in existing_map.keys() {
-        if let Some(item) = new_map.get(id) {
-            if item.current_durability == 0 {
-                items_to_delete.push(id.clone());
-            }
-        } else {
-            items_to_delete.push(id.clone());
-        }
-    }
-
-    // Find items to update (in DB and in new items with different values)
-    let mut items_to_update = Vec::new();
-    for (id, item) in &new_map {
-        if item.current_durability > 0 {
-            if let Some(existing) = existing_map.get(id) {
-                if existing != item {
-                    items_to_update.push(item.clone());
-                }
-            } else {
-                items_to_update.push(item.clone());
-            }
-        }
-    }
-
-    // Delete existing relations - do this once for all items
-    db.query("DELETE FROM owns WHERE in = $owner")
-        .bind(("owner", owner.clone()))
-        .await
-        .map_err(|e| AppError::InternalServerError(format!("Failed to delete items: {}", e)))?;
-
-    // Batch delete operations
-    if !items_to_delete.is_empty() {
-        let delete_ids: Vec<String> = items_to_delete
-            .iter()
-            .map(|id| format!("item:{}", id))
-            .collect();
-
-        db.query("DELETE item WHERE id IN $ids")
-            .bind(("ids", delete_ids))
-            .await
-            .map_err(|e| {
-                AppError::InternalServerError(format!("Failed to batch delete items: {}", e))
-            })?;
-    }
-
-    // Batch update operations
-    if !items_to_update.is_empty() {
-        // Use serde_json::to_value + bound CONTENT so ALL Item fields round-trip
-        // (the previous hand-rolled CONTENT block silently dropped `rarity`,
-        // and string-interpolating fields is fragile around quoting).
-        for item in &items_to_update {
-            let rid = RecordId::from(("item", item.identifier.to_string().as_str()));
-            let body = serde_json::to_value(item).map_err(|e| {
-                AppError::InternalServerError(format!("Failed to encode item: {}", e))
-            })?;
-            db.query("UPSERT $rid CONTENT $body")
-                .bind(("rid", rid))
-                .bind(("body", body))
-                .await
-                .map_err(|e| {
-                    AppError::InternalServerError(format!("Failed to update item: {}", e))
-                })?;
-        }
-
-        // Batch insert relations. Hyphenated UUIDs must be wrapped in
-        // ⟨angle brackets⟩ or Surreal's SQL parser splits them on `-`.
-        let mut relation_parts = Vec::new();
-        for item in &items_to_update {
-            relation_parts.push(format!(
-                "RELATE {}->owns->item:⟨{}⟩",
-                owner, item.identifier
-            ));
-        }
-
-        let bulk_relations = relation_parts.join(";\n");
-        db.query(&bulk_relations).await.map_err(|e| {
-            AppError::InternalServerError(format!("Failed to batch create relations: {}", e))
-        })?;
-    }
-
-    Ok(())
 }
